synthwave-radio-dj/backend/requirements.txt
flask-cors
Flask
requests
elevenlabs==1.5.0
openai>=1.0
python-dotenv

